moved out requests code to a client
