+++ .mise.toml
@@ -1,3 +1,5 @@
+env_file = '.envrc'
+
 [tools]
 uv = 'latest'
 python = 'latest'
